+++ src/components/Tokens/TokenTable/TokenRow/TokenRowCell.tsx
@@ -1,0 +1,129 @@
+import {
+  LARGE_MEDIA_BREAKPOINT,
+  MAX_WIDTH_MEDIA_BREAKPOINT,
+  MEDIUM_MEDIA_BREAKPOINT,
+  SMALL_MEDIA_BREAKPOINT,
+} from 'components/Tokens/constants'
+import styled, { css } from 'styled-components/macro'
+
+const Cell = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`
+export const FavoriteCell = styled(Cell)`
+  min-width: 40px;
+  color: ${({ theme }) => theme.textSecondary};
+  fill: none;
+
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    display: none;
+  }
+`
+export const ListNumberCell = styled(Cell)`
+  color: ${({ theme }) => theme.textSecondary};
+  min-width: 32px;
+  font-size: 14px;
+
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    display: none;
+  }
+`
+export const DataCell = styled(Cell)<{ sortable?: boolean }>`
+  justify-content: flex-end;
+  min-width: 80px;
+  user-select: ${({ sortable }) => (sortable ? 'none' : 'unset')};
+
+  transition: ${({
+    theme: {
+      transition: { duration, timing },
+    },
+  }) => css`background-color ${duration.medium} ${timing.ease}`};
+`
+export const TvlCell = styled(DataCell)`
+  padding-right: 8px;
+  @media only screen and (max-width: ${MEDIUM_MEDIA_BREAKPOINT}) {
+    display: none;
+  }
+`
+export const NameCell = styled(Cell)`
+  justify-content: flex-start;
+  padding: 0px 8px;
+  min-width: 240px;
+  gap: 8px;
+`
+export const PriceCell = styled(DataCell)`
+  padding-right: 8px;
+`
+export const PercentChangeCell = styled(DataCell)`
+  padding-right: 8px;
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    display: none;
+  }
+`
+export const PercentChangeInfoCell = styled(Cell)`
+  display: none;
+
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    display: flex;
+    justify-content: flex-end;
+    color: ${({ theme }) => theme.textSecondary};
+    font-size: 12px;
+    line-height: 16px;
+  }
+`
+export const PriceInfoCell = styled(Cell)`
+  justify-content: flex-end;
+  flex: 1;
+
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    flex-direction: column;
+    align-items: flex-end;
+  }
+`
+export const SparkLineCell = styled(Cell)`
+  padding: 0px 24px;
+  min-width: 120px;
+
+  @media only screen and (max-width: ${MAX_WIDTH_MEDIA_BREAKPOINT}) {
+    display: none;
+  }
+`
+export const SparkLineWrapper = styled(Cell)`
+  width: 124px;
+  height: 42px;
+`
+export const TokenInfoCell = styled(Cell)`
+  gap: 8px;
+  line-height: 24px;
+  font-size: 16px;
+  max-width: inherit;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    justify-content: flex-start;
+    flex-direction: column;
+    gap: 0px;
+    width: max-content;
+    font-weight: 500;
+  }
+`
+export const TokenSymbol = styled(Cell)`
+  color: ${({ theme }) => theme.textTertiary};
+  text-transform: uppercase;
+
+  @media only screen and (max-width: ${SMALL_MEDIA_BREAKPOINT}) {
+    font-size: 12px;
+    height: 16px;
+    justify-content: flex-start;
+    width: 100%;
+  }
+`
+export const VolumeCell = styled(DataCell)`
+  padding-right: 8px;
+  @media only screen and (max-width: ${LARGE_MEDIA_BREAKPOINT}) {
+    display: none;
+  }
+`
